# contract/openapi.json
{
  "components": {
    "schemas": {
      "ErrorDetail": {
        "description": "El cuerpo de `error` que arma build_error_response().",
        "properties": {
          "code": {
            "description": "Codigo estable, snake_case, legible por maquina. Es EL contrato con el frontend. Valores actuales: not_found, conflict, unauthorized, forbidden, validation_error, rate_limited, internal_error, bad_gateway, service_unavailable, bad_request, method_not_allowed, not_acceptable, unsupported_media_type, http_error.",
            "examples": [
              "not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "Contexto estructurado del error. {} cuando no hay nada que anadir.",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Texto para humanos. Puede cambiar sin romper a nadie.",
            "examples": [
              "Resource not found"
            ],
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Id de la peticion, para cruzar con los logs. Ausente cuando la peticion no llego a tener uno.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "El unico formato de error de la API.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Respuesta del healthcheck.",
        "properties": {
          "status": {
            "description": "Estado del servicio. 'ok' cuando responde con normalidad.",
            "examples": [
              "ok"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "RootResponse": {
        "description": "Saludo de la raiz: sirve para confirmar que la API esta arriba.",
        "properties": {
          "message": {
            "examples": [
              "Flash Research API is running"
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "RootResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Flash Research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "getRoot",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Root"
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Health Check"
      }
    }
  }
}